Improved bakeCommand function
In this update you can specify some default values for needed tags and the bakeCommand function will put that default dictionary and then update things to it based on what user entered, e.g. tag "-p" is "." by default for the detectDir command.

The way of running the selected program to run is also changed